pyproject.toml
[tool.poetry]
name = "ufw-monitor-logs"
version = "0.1.1"
description = "Monitor UFW logs and store them in a PostgreSQL database"
authors = ["Hassan El-Masri <hassan@unixtime.com>"]
readme = "README.md"
packages = [{include = "ufw_monitor_logs"}]

[tool.poetry.dependencies]
python = "^3.11"
maxminddb = "^2.4.0"
psycopg2-binary = "^2.9.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
